pyproject.toml
[project]
name = "django_hip_service"
version = "4.0.1"
description = "医院信息平台后台管理系统"
authors = [
    {name = "liaozhiming", email = "liaozhimingandy@qq.com"},
]
dependencies = [
    "openpyxl>=3.1.2",
    "xlsxwriter>=3.2.0",
    "gunicorn>=23.0.0",
    "docutils>=0.20.1",
    "pypinyin>=0.51.0",
    "django-ninja>=1.1.0",
    "pyjwt>=2.8.0",
    "requests>=2.32.2",
    "json2xml>=5.0.1",
    "loguru>=0.7.2",
    "pycryptodome>=3.20.0",
    "gmssl>=3.2.2",
    "django-json-widget>=2.0.1",
    "graphene-django>=3.2.2",
    "django-cors-headers>=4.5.0",
    "python-dotenv>=1.0.1",
    "lxml>=5.3.0",
    "django>=5.1.3",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pyodbc>=5.2.0",
    "python-docx>=1.1.2",
    "psycopg[binary]>=3.2.3",
    "pyyaml>=6.0.2",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "APACHE-2.0"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

test = [
    "pytest>=8.1.1",
]

doc = ["mkdocs"]
[tool.pdm.dev-dependencies]
dev = [
    "django-debug-toolbar>=4.4.6",
]
